Re-use class to mock
